Core: SerializeRaw now has 64-bit size
git-svn-id: svn://ultimatepp.org/upp/trunk@9788 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/Cpu.cpp
@@ -157,11 +157,11 @@ void PokeI64(void *ptr, int64 value) {
 #endif
 #define ENDIAN_SWAP { while(count--) { EndianSwap(*v++); } }
 
-void EndianSwap(word *v, int count) ENDIAN_SWAP
-void EndianSwap(int16 *v, int count) ENDIAN_SWAP
-void EndianSwap(dword *v, int count) ENDIAN_SWAP
-void EndianSwap(int *v, int count) ENDIAN_SWAP
-void EndianSwap(int64 *v, int count) ENDIAN_SWAP
-void EndianSwap(uint64 *v, int count) ENDIAN_SWAP
+void EndianSwap(word *v, size_t count) ENDIAN_SWAP
+void EndianSwap(int16 *v, size_t count) ENDIAN_SWAP
+void EndianSwap(dword *v, size_t count) ENDIAN_SWAP
+void EndianSwap(int *v, size_t count) ENDIAN_SWAP
+void EndianSwap(int64 *v, size_t count) ENDIAN_SWAP
+void EndianSwap(uint64 *v, size_t count) ENDIAN_SWAP
 
 END_UPP_NAMESPACE
